+++ core/vm/actor/builtin/states/storage_power/bind_states.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "vm/actor/builtin/types/type_manager/universal_impl.hpp"
+#include "vm/actor/builtin/types/universal/universal_impl.hpp"
 
 #include "vm/actor/builtin/states/storage_power/storage_power_actor_state.hpp"
 #include "vm/actor/builtin/states/storage_power/v0/storage_power_actor_state.hpp"
