+++ core/vm/actor/builtin/states/storage_power/storage_power_actor_state.hpp
@@ -14,7 +14,7 @@
 #include "primitives/types.hpp"
 #include "vm/actor/builtin/types/storage_power/claim.hpp"
 #include "vm/actor/builtin/types/storage_power/cron_event.hpp"
-#include "vm/actor/builtin/types/type_manager/universal.hpp"
+#include "vm/actor/builtin/types/universal/universal.hpp"
 
 // Forward declaration
 namespace fc::vm::runtime {
