+++ core/vm/actor/builtin/states/system/system_actor_state.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "vm/actor/builtin/types/type_manager/universal.hpp"
+#include "vm/actor/builtin/types/universal/universal.hpp"
 
 namespace fc::vm::actor::builtin::states {
