+++ core/vm/actor/builtin/types/miner/bind_types.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "vm/actor/builtin/types/type_manager/universal_impl.hpp"
+#include "vm/actor/builtin/types/universal/universal_impl.hpp"
 
 #include "vm/actor/builtin/types/miner/v0/expiration.hpp"
 #include "vm/actor/builtin/types/miner/v2/expiration.hpp"
